+++ rust/shekyl-archival-retention/src/lib.rs
@@ -25,6 +25,7 @@
 
 #![deny(unsafe_code)]
 
+pub mod bond_connect;
 pub mod bond_ct_balance;
 pub mod bond_floor;
 pub mod bond_post;
@@ -49,6 +50,10 @@ pub mod serve_credit_decisions;
 pub mod serve_eligibility;
 pub mod wire;
 
+pub use bond_connect::{
+    clean_interval_close, is_clean_interval_close, unbond_connect, unbond_pop, UnbondConnect,
+    UnbondConnectError, UnbondPopError, MAX_BOND_BAD_INTERVALS,
+};
 pub use bond_ct_balance::{verify_bond_post_ct_balance, BondCtBalanceError, BondTerm};
 pub use bond_floor::{
     bond_floor, ARCHIVAL_BOND_FLOOR_ATOMIC, ARCHIVAL_REORG_DEPTH_BLOCKS,